requirements.txt
discord-py-interactions>=4.3.4
interactions-persistence
interactions-lavalink
interactions-tasks
interactions-files
pet-pet-gif
loguru
pytz
python-decouple
aiofiles
aiosqlite
qrcode[pil]
motor[srv]
